feat: add an internal for now studio package (#11037)
* feat: add an internal for now studio package

* fix: remove test commands

* fix: use getenvs from studio package

* fix: packagejson

* chore: changeset

* fix: remove unused error

* fix: of course Bjorn was right all along
