+++ .changeset/silent-socks-return.md
@@ -1,0 +1,5 @@
+---
+"@astrojs/db": patch
+---
+
+Internal refactor, this change should have no visible effect
