+++ .changeset/young-plums-doubt.md
@@ -1,0 +1,5 @@
+---
+"@astrojs/studio": minor
+---
+
+Created package. This package contain shared code between integrations that interact with Astro Studio and is not intended to be used by end-users at this time
