+++ packages/db/src/core/cli/commands/push/index.ts
@@ -3,7 +3,7 @@ import prompts from 'prompts';
 import type { Arguments } from 'yargs-parser';
 import { safeFetch } from '../../../../runtime/utils.js';
 import { MIGRATION_VERSION } from '../../../consts.js';
-import { getManagedAppTokenOrExit } from '../../../tokens.js';
+import { getManagedAppTokenOrExit } from '@astrojs/studio';
 import { type DBConfig, type DBSnapshot } from '../../../types.js';
 import { type Result, getRemoteDatabaseUrl } from '../../../utils.js';
 import {
