+++ packages/db/src/core/cli/commands/verify/index.ts
@@ -1,6 +1,6 @@
 import type { AstroConfig } from 'astro';
 import type { Arguments } from 'yargs-parser';
-import { getManagedAppTokenOrExit } from '../../../tokens.js';
+import { getManagedAppTokenOrExit } from '@astrojs/studio';
 import type { DBConfig } from '../../../types.js';
 import {
 	createCurrentSnapshot,
