+++ packages/studio/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@astrojs/studio",
+  "version": "0.0.1",
+  "description": "Internal package powering integrations between Astro projects and Astro Studio",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/withastro/astro.git",
+    "directory": "packages/studio"
+  },
+  "bugs": "https://github.com/withastro/astro/issues",
+  "homepage": "https://docs.astro.build/en/guides/integrations-guide/studio/",
+  "type": "module",
+  "author": "withastro",
+  "types": "./dist/index.js",
+  "main": "./dist/index.js",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "files": [
+    "dist"
+  ],
+  "keywords": [
+    "withastro",
+    "astro-integration"
+  ],
+  "scripts": {
+    "build": "astro-scripts build \"src/**/*.ts\" && tsc",
+    "build:ci": "astro-scripts build \"src/**/*.ts\"",
+    "dev": "astro-scripts dev \"src/**/*.ts\""
+  },
+  "dependencies": {
+    "ci-info": "^4.0.0",
+    "kleur": "^4.1.5",
+    "ora": "^8.0.1"
+  },
+  "devDependencies": {
+    "astro": "workspace:*",
+    "astro-scripts": "workspace:*",
+    "typescript": "^5.4.5",
+    "vite": "^5.2.11"
+  }
+}
